crypto: Breakdown crypto.rs in sui-framework
