+++ crates/sui-framework/docs/bulletproofs.md
@@ -5,26 +5,11 @@
 
 
 
--  [Constants](#@Constants_0)
 -  [Function `native_verify_full_range_proof`](#0x2_bulletproofs_native_verify_full_range_proof)
 -  [Function `verify_full_range_proof`](#0x2_bulletproofs_verify_full_range_proof)
 
 
 <pre><code><b>use</b> <a href="elliptic_curve.md#0x2_elliptic_curve">0x2::elliptic_curve</a>;
-</code></pre>
-
-
-
-<a name="@Constants_0"></a>
-
-## Constants
-
-
-<a name="0x2_bulletproofs_EBulletproofsVerificationFailed"></a>
-
-
-
-<pre><code><b>const</b> <a href="bulletproofs.md#0x2_bulletproofs_EBulletproofsVerificationFailed">EBulletproofsVerificationFailed</a>: u64 = 5;
 </code></pre>
 
 
@@ -36,7 +21,7 @@
 Only bit_length = 64, 32, 16, 8 will work.
 
 
-<pre><code><b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_native_verify_full_range_proof">native_verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &<a href="">vector</a>&lt;u8&gt;, bit_length: u64)
+<pre><code><b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_native_verify_full_range_proof">native_verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &<a href="">vector</a>&lt;u8&gt;, bit_length: u64): bool
 </code></pre>
 
 
@@ -45,7 +30,7 @@ Only bit_length = 64, 32, 16, 8 will work.
 <summary>Implementation</summary>
 
 
-<pre><code><b>native</b> <b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_native_verify_full_range_proof">native_verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &<a href="">vector</a>&lt;u8&gt;, bit_length: u64);
+<pre><code><b>native</b> <b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_native_verify_full_range_proof">native_verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &<a href="">vector</a>&lt;u8&gt;, bit_length: u64): bool;
 </code></pre>
 
 
@@ -63,7 +48,7 @@ Only bit_length = 64, 32, 16, 8 will work.
 If the range proof is valid, execution succeeds, else panics.
 
 
-<pre><code><b>public</b> <b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_verify_full_range_proof">verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &<a href="elliptic_curve.md#0x2_elliptic_curve_RistrettoPoint">elliptic_curve::RistrettoPoint</a>, bit_length: u64)
+<pre><code><b>public</b> <b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_verify_full_range_proof">verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &<a href="elliptic_curve.md#0x2_elliptic_curve_RistrettoPoint">elliptic_curve::RistrettoPoint</a>, bit_length: u64): bool
 </code></pre>
 
 
@@ -72,7 +57,7 @@ If the range proof is valid, execution succeeds, else panics.
 <summary>Implementation</summary>
 
 
-<pre><code><b>public</b> <b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_verify_full_range_proof">verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &RistrettoPoint, bit_length: u64) {
+<pre><code><b>public</b> <b>fun</b> <a href="bulletproofs.md#0x2_bulletproofs_verify_full_range_proof">verify_full_range_proof</a>(proof: &<a href="">vector</a>&lt;u8&gt;, commitment: &RistrettoPoint, bit_length: u64): bool {
     <a href="bulletproofs.md#0x2_bulletproofs_native_verify_full_range_proof">native_verify_full_range_proof</a>(proof, &ec::bytes(commitment), bit_length)
 }
 </code></pre>
