+++ crates/sui-framework/src/natives/crypto/ed25519.rs
@@ -1,0 +1,51 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+use crate::legacy_empty_cost;
+use fastcrypto::{
+    ed25519::{Ed25519PublicKey, Ed25519Signature},
+    traits::ToFromBytes,
+    Verifier,
+};
+use move_binary_format::errors::PartialVMResult;
+use move_vm_runtime::native_functions::NativeContext;
+use move_vm_types::{
+    loaded_data::runtime_types::Type,
+    natives::function::NativeResult,
+    pop_arg,
+    values::{Value, VectorRef},
+};
+use smallvec::smallvec;
+use std::collections::VecDeque;
+
+pub fn ed25519_verify(
+    _context: &mut NativeContext,
+    ty_args: Vec<Type>,
+    mut args: VecDeque<Value>,
+) -> PartialVMResult<NativeResult> {
+    debug_assert!(ty_args.is_empty());
+    debug_assert!(args.len() == 3);
+
+    let msg = pop_arg!(args, VectorRef);
+    let msg_ref = msg.as_bytes_ref();
+    let public_key_bytes = pop_arg!(args, VectorRef);
+    let public_key_bytes_ref = public_key_bytes.as_bytes_ref();
+    let signature_bytes = pop_arg!(args, VectorRef);
+    let signature_bytes_ref = signature_bytes.as_bytes_ref();
+
+    let cost = legacy_empty_cost();
+
+    let signature = match <Ed25519Signature as ToFromBytes>::from_bytes(&signature_bytes_ref) {
+        Ok(signature) => signature,
+        Err(_) => return Ok(NativeResult::ok(cost, smallvec![Value::bool(false)])),
+    };
+
+    let public_key = match <Ed25519PublicKey as ToFromBytes>::from_bytes(&public_key_bytes_ref) {
+        Ok(public_key) => public_key,
+        Err(_) => return Ok(NativeResult::ok(cost, smallvec![Value::bool(false)])),
+    };
+
+    match public_key.verify(&msg_ref, &signature) {
+        Ok(_) => Ok(NativeResult::ok(cost, smallvec![Value::bool(true)])),
+        Err(_) => Ok(NativeResult::ok(cost, smallvec![Value::bool(false)])),
+    }
+}
