+++ crates/sui-framework/src/natives/crypto/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod bls12381;
+pub mod bulletproofs;
+pub mod ecdsa_k1;
+pub mod ed25519;
+pub mod elliptic_curve;
+pub mod groth16;
+pub mod hmac;
